Remove all "upcoming assignments" stuff
I don't think it's possible to do without violating the standalone modularity of the datasources. I haven't found a way to get all unit IDs at once from the CloudDeakin API, so to query them all for all their assignments in one go (requiring all their IDs to be available) would currently require calling the DeakinSync API. Also this is probably not an essential feature, everything's likely to be grouped by unit anyway or can be otherwise grouped after retrieval on the front-end.